ajuste das regras

+++ MainClass.java
@@ -2,11 +2,11 @@ import java.util.Scanner;
 public class MainClass{ 
   public static void main(String args[]){
        Scanner _key = new Scanner(System.in);
-double  a;
-double  b;
+int  a;
+int  b;
 String  t1;
-a= _key.nextDouble();
-b= _key.nextDouble();
+a= _key.nextInt();
+b= _key.nextInt();
 a = 1+2*3/b;
 if (a>b) {
 System.out.println(a);}else {
